분석결과 정리, 식단 날짜 검증로직 추가

+++ main_project/src/main/java/NutrientsCoders/main_project/food/dto/FoodPatchDto.java
@@ -8,7 +8,7 @@ public class FoodPatchDto {
   
   private Long servingSize;
   
-  private Long kcal;
+  private Double kcal;
   
   private Double carbo;
